chore!: move discovery-related fns/views to discovery module add reusable data type summary functions change overall overview shape to use data type summary fns
