+++ chord_metadata_service/chord/views_search.py
@@ -1,12 +1,16 @@
+import asyncio
 import itertools
 import json
 
+from adrf.decorators import api_view as async_api_view
+from bento_lib.auth.permissions import P_QUERY_DATASET_LEVEL_COUNTS
+from bento_lib.auth.resources import build_resource
 from bento_lib.responses import errors
 from bento_lib.search import build_search_response, postgres
 
 from datetime import datetime
 from django.db import connection
-from django.db.models import Count, F, Q
+from django.db.models import Count, F, Q, QuerySet
 from django.db.models.functions import Coalesce
 from django.contrib.postgres.aggregates import ArrayAgg
 from django.conf import settings
@@ -22,74 +26,52 @@ from typing import Callable
 from chord_metadata_service.authz.middleware import authz_middleware
 
 from chord_metadata_service.logger import logger
-from chord_metadata_service.restapi.utils import queryset_stats_for_field
 
 from chord_metadata_service.experiments.api_views import EXPERIMENT_SELECT_REL, EXPERIMENT_PREFETCH
 from chord_metadata_service.experiments.models import Experiment
 from chord_metadata_service.experiments.serializers import ExperimentSerializer
 
-
 from chord_metadata_service.metadata.elastic import es
-
-from chord_metadata_service.patients.models import Individual
 
 from chord_metadata_service.phenopackets.api_views import PHENOPACKET_SELECT_REL, PHENOPACKET_PREFETCH
 from chord_metadata_service.phenopackets.models import Phenopacket, Biosample
 from chord_metadata_service.phenopackets.serializers import PhenopacketSerializer
+from chord_metadata_service.phenopackets.summaries import dt_phenopacket_summary
 
 from .data_types import DATA_TYPE_EXPERIMENT, DATA_TYPE_PHENOPACKET, DATA_TYPES
 from .models import Dataset
 
 from collections import defaultdict
 
+from ..authz.discovery import get_data_type_discovery_permissions
+from ..experiments.summaries import dt_experiment_summary
 
 OUTPUT_FORMAT_VALUES_LIST = "values_list"
 OUTPUT_FORMAT_BENTO_SEARCH_RESULT = "bento_search_result"
 
 
-def experiment_dataset_summary(dataset):
-    # TODO; deduplicate with chord_metadata_service.restapi
+async def experiment_dataset_summary(request: Request, dataset):
+    return await dt_experiment_summary(
+        Experiment.objects.filter(dataset=dataset),
+        dt_permissions=await get_data_type_discovery_permissions(
+            request,
+            [DATA_TYPE_EXPERIMENT],
+            build_resource(project=dataset.project_id, dataset=dataset.identifier),
+            counts_permission=P_QUERY_DATASET_LEVEL_COUNTS,
+        )
+    )
 
-    experiments = Experiment.objects.filter(dataset=dataset)
 
-    return {
-        "count": experiments.count(),
-        "data_type_specific": {},  # TODO
-    }
-
-
-async def phenopacket_dataset_summary(dataset):
-    # TODO; deduplicate with chord_metadata_service.restapi
-    phenopacket_qs = Phenopacket.objects.filter(dataset=dataset)  # TODO
-
-    # Sex related fields stats are precomputed here and post processed later
-    # to include missing values inferred from the schema
-    individuals_sex = await queryset_stats_for_field(phenopacket_qs, "subject__sex")
-    individuals_k_sex = await queryset_stats_for_field(phenopacket_qs, "subject__karyotypic_sex")
-
-    return {
-        "count": phenopacket_qs.count(),
-        "data_type_specific": {
-            # TODO: deduplicate with biosamples summary from REST API
-            "biosamples": {
-                "count": phenopacket_qs.values("biosamples__id").count(),
-                "is_control_sample": queryset_stats_for_field(phenopacket_qs, "biosamples__is_control_sample"),
-                "taxonomy": queryset_stats_for_field(phenopacket_qs, "biosamples__taxonomy__label"),
-            },
-            # TODO: deduplicate with diseases summary from REST API
-            "diseases": queryset_stats_for_field(phenopacket_qs, "diseases__term__label"),
-            # TODO: deduplicate with individuals summary from REST API
-            "individuals": {
-                "count": phenopacket_qs.values("subject__id").count(),
-                "sex": {k: individuals_sex.get(k, 0) for k in (s[0] for s in Individual.SEX)},
-                "karyotypic_sex": {k: individuals_k_sex.get(k, 0) for k in (s[0] for s in Individual.KARYOTYPIC_SEX)},
-                "taxonomy": queryset_stats_for_field(phenopacket_qs, "subject__taxonomy__label"),
-                "date_of_birth": phenopacket_qs.values("subject__date_of_birth")
-            },
-            # TODO: deduplicate with phenotypic features summary from REST API
-            "phenotypic_features": queryset_stats_for_field(phenopacket_qs, "phenotypic_features__pftype__label"),
-        }
-    }
+async def phenopacket_dataset_summary(request: Request, dataset: Dataset):
+    return await dt_phenopacket_summary(
+        Phenopacket.objects.filter(dataset=dataset),
+        dt_permissions=await get_data_type_discovery_permissions(
+            request,
+            [DATA_TYPE_PHENOPACKET],
+            build_resource(project=dataset.project_id, dataset=dataset.identifier),
+            counts_permission=P_QUERY_DATASET_LEVEL_COUNTS,
+        )
+    )
 
 
 def get_field_lookup(field: list[str]) -> str:
@@ -118,14 +100,14 @@ def get_values_list(queryset, options):
     return queryset.values_list(field_lookup, flat=True)
 
 
-def data_type_results(query, params, key="id"):
+def data_type_results(query: sql.SQL, params, key="id"):
     with connection.cursor() as cursor:
         logger.debug(f"Executing SQL:\n    {query.as_string(cursor.connection)}")
         cursor.execute(query.as_string(cursor.connection), params)
         return set(dict(zip([col[0] for col in cursor.description], row))[key] for row in cursor.fetchall())
 
 
-def experiment_query_results(query, params, options=None):
+def experiment_query_results(query: sql.SQL, params, options=None) -> QuerySet:
     # TODO: possibly a quite inefficient way of doing things...
     # TODO: Prefetch related biosample or no?
     queryset = Experiment.objects.filter(id__in=data_type_results(query, params, "id"))
@@ -134,11 +116,14 @@ def experiment_query_results(query, params, options=None):
     if output_format == OUTPUT_FORMAT_VALUES_LIST:
         return get_values_list(queryset, options)
 
-    return queryset.select_related(*EXPERIMENT_SELECT_REL) \
+    return (
+        queryset
+        .select_related(*EXPERIMENT_SELECT_REL)
         .prefetch_related(*EXPERIMENT_PREFETCH)
+    )
 
 
-def get_biosamples_with_experiment_details(subject_ids):
+def get_biosamples_with_experiment_details(subject_ids) -> QuerySet:
     """
     The function returns a queryset where each entry represents a biosample obtained from a subject, along with
     details of any associated experiment. If a biosample does not have an associated experiment, the experiment
@@ -181,10 +166,11 @@ def phenopacket_query_results(query, params, options=None):
         )
 
         # Get the biosamples with experiments data
-        phenopacket_ids = [result['subject_id'] for result in results]
+        phenopacket_ids = [result["subject_id"] for result in results]
         biosamples_experiments_details = get_biosamples_with_experiment_details(phenopacket_ids)
 
         # Group the experiments with biosamples by subject_id
+        # TODO: this is a hack, and also hard-codes the biosample relationship. Redo it.
         experiments_with_biosamples = defaultdict(list)
         for b in biosamples_experiments_details:
             experiments_with_biosamples[b["subject_id"]].append({
@@ -202,12 +188,16 @@ def phenopacket_query_results(query, params, options=None):
 
         # Add the experiments_with_biosamples data to the results
         for result in results:
-            result["experiments_with_biosamples"] = experiments_with_biosamples[result['subject_id']]
+            result["experiments_with_biosamples"] = experiments_with_biosamples[result["subject_id"]]
 
         return results
-    else:
-        return queryset.select_related(*PHENOPACKET_SELECT_REL) \
-            .prefetch_related(*PHENOPACKET_PREFETCH)
+
+    # Otherwise, old format
+    return (
+        queryset
+        .select_related(*PHENOPACKET_SELECT_REL)
+        .prefetch_related(*PHENOPACKET_PREFETCH)
+    )
 
 
 QUERY_RESULTS_FN: dict[str, Callable] = {
@@ -252,7 +242,8 @@ def search(request, internal_data=False):
             params=query_params,
             key="dataset_id"
         ))
-        return Response(build_search_response([{"id": d.identifier, "data_type": data_type} for d in datasets], start))
+        return Response(
+            build_search_response([{"id": d.identifier, "data_type": data_type} for d in datasets], start))
 
     serializer_class = QUERY_RESULT_SERIALIZERS[data_type]
     query_function = QUERY_RESULTS_FN[data_type]
@@ -582,11 +573,20 @@ def private_dataset_search(request: Request, dataset_id: str):
     return dataset_search(request, dataset_id, internal=True)
 
 
-@api_view(["GET"])
-def dataset_summary(request: Request, dataset_id: str):
+DATASET_DATA_TYPE_SUMMARY_FUNCTIONS = {
+    DATA_TYPE_PHENOPACKET: phenopacket_dataset_summary,
+    DATA_TYPE_EXPERIMENT: experiment_dataset_summary,
+}
+
+
+@async_api_view(["GET"])
+async def dataset_summary(request: Request, dataset_id: str):
     # TODO: PERMISSIONS
     dataset = Dataset.objects.get(identifier=dataset_id)
-    return Response({
-        DATA_TYPE_PHENOPACKET: phenopacket_dataset_summary(dataset=dataset),
-        DATA_TYPE_EXPERIMENT: experiment_dataset_summary(dataset=dataset),
-    })
+
+    summaries = await asyncio.gather(
+        *map(lambda dt: DATASET_DATA_TYPE_SUMMARY_FUNCTIONS[dt](request, dataset),
+             DATASET_DATA_TYPE_SUMMARY_FUNCTIONS.keys())
+    )
+
+    return Response({dt: s} for dt, s in zip(DATASET_DATA_TYPE_SUMMARY_FUNCTIONS.keys(), summaries))
